# New method of cleaning databases to improve test perf