--- a/test/Microsoft.EntityFrameworkCore.InMemory.FunctionalTests/NullKeysInMemoryTest.cs
+++ b/test/Microsoft.EntityFrameworkCore.InMemory.FunctionalTests/NullKeysInMemoryTest.cs
@@ -33,6 +33,12 @@ namespace Microsoft.EntityFrameworkCore.InMemory.FunctionalTests
 
             public override DbContext CreateContext()
                 => new DbContext(_options);
+
+            protected override void EnsureClean(DbContext context)
+            {
+                context.Database.EnsureDeleted();
+                context.Database.EnsureCreated();
+            }
         }
     }
 }