--- a/test/Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests/CompositeKeyEndToEndTest.cs
+++ b/test/Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests/CompositeKeyEndToEndTest.cs
@@ -24,7 +24,7 @@ namespace Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests
 
             using (var context = new BronieContext(serviceProvider, "CompositePegasuses"))
             {
-                context.Database.EnsureCreated();
+                context.Database.EnsureClean();
 
                 context.Add(new Pegasus { Id1 = ticks, Id2 = ticks + 1, Name = "Rainbow Dash" });
                 await context.SaveChangesAsync();
@@ -69,7 +69,7 @@ namespace Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests
 
             using (var context = new BronieContext(serviceProvider, "CompositeUnicorns"))
             {
-                context.Database.EnsureCreated();
+                context.Database.EnsureClean();
 
                 var added = context.Add(new Unicorn { Id2 = id2, Name = "Rarity" }).Entity;
 
@@ -126,7 +126,7 @@ namespace Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests
 
             using (var context = new BronieContext(serviceProvider, "CompositeEarthPonies"))
             {
-                context.Database.EnsureCreated();
+                context.Database.EnsureClean();
 
                 var pony1 = context.Add(new EarthPony { Id2 = 7, Name = "Apple Jack 1" }).Entity;
                 var pony2 = context.Add(new EarthPony { Id2 = 7, Name = "Apple Jack 2" }).Entity;