--- a/test/Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests/DataAnnotationSqlServerFixture.cs
+++ b/test/Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests/DataAnnotationSqlServerFixture.cs
@@ -62,12 +62,8 @@ namespace Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests
 
                     using (var context = new DataAnnotationContext(optionsBuilder.Options))
                     {
-                        // TODO: Delete DB if model changed
-                        context.Database.EnsureDeleted();
-                        if (context.Database.EnsureCreated())
-                        {
-                            DataAnnotationModelInitializer.Seed(context);
-                        }
+                        context.Database.EnsureClean();
+                        DataAnnotationModelInitializer.Seed(context);
 
                         TestSqlLoggerFactory.Reset();
                     }