… src/Microsoft.EntityFrameworkCore.Relational.Specification.Tests/project.json
… test/Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests/FindSqlServerTest.cs
… test/Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests/GearsOfWarQuerySqlServerFixture.cs
… test/Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests/GraphUpdatesSqlServerTestBase.cs
… test/Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests/InheritanceRelationshipsQuerySqlServerFixture.cs
… test/Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests/InheritanceSqlServerFixture.cs
… test/Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests/Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests.csproj
… test/Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests/MonsterFixupSqlServerTest.cs
… test/Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests/NotificationEntitiesSqlServerTest.cs
… test/Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests/NullKeysSqlServerTest.cs
… test/Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests/NullSemanticsQuerySqlServerFixture.cs
… test/Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests/OneToOneQuerySqlServerFixture.cs
… test/Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests/QueryBugsTest.cs
… test/Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests/SequenceEndToEndTest.cs
… test/Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests/SequentialGuidEndToEndTest.cs
… test/Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests/SqlServerDatabaseCleaner.cs
… test/Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests/SqlServerDatabaseFacadeExtensions.cs
… test/Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests/SqlServerEndToEndTest.cs
… test/Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests/SqlServerValueGenerationScenariosTest.cs
… test/Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests/StoreGeneratedSqlServerTest.cs
… test/Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests/UpdatesSqlServerFixture.cs
… test/Microsoft.EntityFrameworkCore.SqlServer.FunctionalTests/project.json
… test/Microsoft.EntityFrameworkCore.Sqlite.FunctionalTests/AutoincrementTest.cs
… test/Microsoft.EntityFrameworkCore.Sqlite.FunctionalTests/BuiltInDataTypesSqliteFixture.cs
… test/Microsoft.EntityFrameworkCore.Sqlite.FunctionalTests/CommandConfigurationTest.cs
… test/Microsoft.EntityFrameworkCore.Sqlite.FunctionalTests/ComplexNavigationsQuerySqliteFixture.cs
… test/Microsoft.EntityFrameworkCore.Sqlite.FunctionalTests/CompositeKeyEndToEndTest.cs
… test/Microsoft.EntityFrameworkCore.Sqlite.FunctionalTests/DataAnnotationSqliteFixture.cs
… test/Microsoft.EntityFrameworkCore.Sqlite.FunctionalTests/DefaultValuesTest.cs
… test/Microsoft.EntityFrameworkCore.Sqlite.FunctionalTests/F1SqliteFixture.cs