@roji 2016-07-12
Wow, nice idea. Npgsql tests will definitely benefit from this as creating a database is a pretty heavy operation in PostgreSQL...